api-azure/requirements.txt
azure-functions
azure-data-tables
llama-index>=0.10.0,<0.15.0
openai>=1.0.0
tiktoken>=0.5.0